Add expenses

+++ common/src/lib.rs
@@ -25,7 +25,7 @@ pub struct Expense {
     pub id: ExpenseId,
     pub payer: UserId,
     pub participants: Vec<UserId>,
-    pub amount: f64,
+    pub amount: u64,
     pub description: Option<String>,
     pub group_id: Option<GroupId>,
     pub timestamp_ms: i64,
